requirements.txt
# ============================================================
# 🎬 ERP NUNCA FILMES - Backend FastAPI
# Python 3.12 / Render.com
# ============================================================

# --- Framework principal ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Banco de dados / ORM ---
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3  # Driver PostgreSQL moderno
sqlite-utils==3.37  # Utilitário útil em dev/testes

# --- Validação e configuração ---
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# --- Segurança e CORS ---
starlette==0.37.2
cryptography==43.0.3

# --- Utilitários gerais ---
requests==2.32.3
python-multipart==0.0.9
typing-extensions==4.12.2

# --- Logs e monitoramento ---
loguru==0.7.3

# --- Migração e ETL ---
pandas==2.2.3
openpyxl==3.1.5

# --- Testes ---
pytest==8.3.3
httpx==0.27.2

# --- Compatibilidade Render / Deploy ---
gunicorn==23.0.0

# ============================================================
# Observações:
# - psycopg[binary] é usado no Render (não requer compilação)
# - alembic controla versões do schema (migrations)
# - python-dotenv e pydantic-settings para .env dinâmico
# - loguru melhora logs de produção
# ============================================================
